[mod]修改了WtCfgLoader自动检测编码，调用的地方都做了对应的调整

+++ src/Includes/WTSDataDef.hpp
@@ -424,7 +424,7 @@ public:
 				vArray->append(day.date);
 				break;
 			case KFT_TIME:
-				vArray->append(day.time);
+				vArray->append((double)day.time);
 			}
 		}
 
@@ -758,7 +758,7 @@ public:
 				vArray->append(day.date);
 				break;
 			case KFT_TIME:
-				vArray->append(day.time);
+				vArray->append((double)day.time);
 			}
 		}
